Remove unused pipeline components and parsers
Deleted multiple files related to KFP components, GCP integration, pipelines, and their corresponding tests. This cleanup focuses on simplifying the codebase by removing redundant or unused functionality.
